+++ src/oemof/solph/models.py
@@ -1,21 +1,26 @@
 # -*- coding: utf-8 -*-
-"""Solph Optimization Models
 
-This file is part of project oemof (github.com/oemof/oemof). It's copyrighted
-by the contributors recorded in the version control history of the file,
-available from its original location oemof/oemof/solph/models.py
+"""Solph Optimization Models.
+
+SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
+SPDX-FileCopyrightText: Simon Hilpert
+SPDX-FileCopyrightText: Cord Kaldemeyer
+SPDX-FileCopyrightText: gplssm
+SPDX-FileCopyrightText: Patrik Schönfeldt
 
 SPDX-License-Identifier: MIT
+
 """
 import logging
 import warnings
 
-import pyomo.environ as po
+from pyomo import environ as po
+from pyomo.core.plugins.transform.relax_integrality import RelaxIntegrality
+from pyomo.opt import SolverFactory
+
 from oemof.solph import blocks
 from oemof.solph import processing
 from oemof.solph.plumbing import sequence
-from pyomo.core.plugins.transform.relax_integrality import RelaxIntegrality
-from pyomo.opt import SolverFactory
 
 
 class BaseModel(po.ConcreteModel):
@@ -28,7 +33,7 @@ class BaseModel(po.ConcreteModel):
     constraint_groups : list (optional)
         Solph looks for these groups in the given energy system and uses them
         to create the constraints of the optimization problem.
-        Defaults to :const:`Model.CONSTRAINTS`
+        Defaults to `Model.CONSTRAINTS`
     objective_weighting : array like (optional)
         Weights used for temporal objective function
         expressions. If nothing is passed `timeincrement` will be used which
@@ -201,21 +206,19 @@ class BaseModel(po.ConcreteModel):
 
         solver_results = opt.solve(self, **solve_kwargs)
 
-        status = solver_results["Solver"][0]["Status"].key
+        status = solver_results["Solver"][0]["Status"]
         termination_condition = (
-            solver_results["Solver"][0]["Termination condition"].key)
+            solver_results["Solver"][0]["Termination condition"])
 
         if status == "ok" and termination_condition == "optimal":
             logging.info("Optimization successful...")
-            self.es.results = solver_results
-            self.solver_results = solver_results
         else:
             msg = ("Optimization ended with status {0} and termination "
                    "condition {1}")
             warnings.warn(msg.format(status, termination_condition),
                           UserWarning)
-            self.es.results = solver_results
-            self.solver_results = solver_results
+        self.es.results = solver_results
+        self.solver_results = solver_results
 
         return solver_results
 
@@ -238,7 +241,7 @@ class Model(BaseModel):
     constraint_groups : list
         Solph looks for these groups in the given energy system and uses them
         to create the constraints of the optimization problem.
-        Defaults to :const:`Model.CONSTRAINTS`
+        Defaults to `Model.CONSTRAINTS`
 
     **The following basic sets are created**:
 
@@ -303,24 +306,28 @@ class Model(BaseModel):
                            within=po.Reals)
 
         for (o, i) in self.FLOWS:
-            for t in self.TIMESTEPS:
-                if (o, i) in self.UNIDIRECTIONAL_FLOWS:
-                    self.flow[o, i, t].setlb(0)
-                if self.flows[o, i].nominal_value is not None:
-                    self.flow[o, i, t].setub(self.flows[o, i].max[t] *
-                                             self.flows[o, i].nominal_value)
-
-                    if self.flows[o, i].actual_value[t] is not None:
-                        # pre- optimized value of flow variable
+            if self.flows[o, i].nominal_value is not None:
+                if self.flows[o, i].fix[self.TIMESTEPS[1]] is not None:
+                    for t in self.TIMESTEPS:
                         self.flow[o, i, t].value = (
-                            self.flows[o, i].actual_value[t] *
+                            self.flows[o, i].fix[t] *
                             self.flows[o, i].nominal_value)
-                        # fix variable if flow is fixed
-                        if self.flows[o, i].fixed:
-                            self.flow[o, i, t].fix()
+                        self.flow[o, i, t].fix()
+                else:
+                    for t in self.TIMESTEPS:
+                        self.flow[o, i, t].setub(
+                            self.flows[o, i].max[t] *
+                            self.flows[o, i].nominal_value)
 
                     if not self.flows[o, i].nonconvex:
-                        # lower bound of flow variable
-                        self.flow[o, i, t].setlb(
-                            self.flows[o, i].min[t] *
-                            self.flows[o, i].nominal_value)
+                        for t in self.TIMESTEPS:
+                            self.flow[o, i, t].setlb(
+                                self.flows[o, i].min[t] *
+                                self.flows[o, i].nominal_value)
+                    elif (o, i) in self.UNIDIRECTIONAL_FLOWS:
+                        for t in self.TIMESTEPS:
+                            self.flow[o, i, t].setlb(0)
+            else:
+                if (o, i) in self.UNIDIRECTIONAL_FLOWS:
+                    for t in self.TIMESTEPS:
+                        self.flow[o, i, t].setlb(0)