+++ tests/test_scripts/test_solph/test_variable_chp/test_variable_chp.py
@@ -17,6 +17,7 @@ import os
 
 import pandas as pd
 from nose.tools import eq_
+
 from oemof import solph
 from oemof.network.network import Node
 from oemof.solph import views
@@ -61,15 +62,15 @@ def test_variable_chp(filename="variable_chp.csv", solver='cbc'):
 
     # create simple sink object for electrical demand for each electrical bus
     solph.Sink(label=('demand', 'elec1'), inputs={bel: solph.Flow(
-        actual_value=data['demand_el'], fixed=True, nominal_value=1)})
+        fix=data['demand_el'], nominal_value=1)})
     solph.Sink(label=('demand', 'elec2'), inputs={bel2: solph.Flow(
-        actual_value=data['demand_el'], fixed=True, nominal_value=1)})
+        fix=data['demand_el'], nominal_value=1)})
 
     # create simple sink object for heat demand for each thermal bus
     solph.Sink(label=('demand', 'therm1'), inputs={bth: solph.Flow(
-        actual_value=data['demand_th'], fixed=True, nominal_value=741000)})
+        fix=data['demand_th'], nominal_value=741000)})
     solph.Sink(label=('demand', 'therm2'), inputs={bth2: solph.Flow(
-        actual_value=data['demand_th'], fixed=True, nominal_value=741000)})
+        fix=data['demand_th'], nominal_value=741000)})
 
     # create a fixed transformer to distribute to the heat_2 and elec_2 buses
     solph.Transformer(
@@ -117,12 +118,12 @@ def test_variable_chp(filename="variable_chp.csv", solver='cbc'):
 
     for key in variable_chp_dict_max.keys():
         logging.debug("Test the maximum value of {0}".format(key))
-        eq_(int(round(maxresults[key])),
+        eq_(int(round(maxresults[[key]])),
             int(round(variable_chp_dict_max[key])))
 
     for key in variable_chp_dict_sum.keys():
         logging.debug("Test the summed up value of {0}".format(key))
-        eq_(int(round(sumresults[key])),
+        eq_(int(round(sumresults[[key]])),
             int(round(variable_chp_dict_sum[key])))
 
     eq_(parameter[(energysystem.groups["('fixed_chp', 'gas')"], None)]